requirements.txt
agno>=2.5.0
a2a-sdk>=0.3.0
openai>=2.0.0
fastapi
uvicorn[standard]
httpx
httpx-sse
python-dotenv
